LayoutTests/platform/chromium/fast/forms/calendar-picker/month-picker-key-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=month id=month value="2000-01">
<script>
description("Tests if month picker key bindings work as expected.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('month'), test1);

function test1() {
    debug('Check that page popup exists.');
    shouldBe('popupWindow.pagePopupController.toString()', '"[object PagePopupController]"');

    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31');
    shouldBeEqualToString('highlightedValue()', '2000-01');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31');

    debug('Check that arrow keys work properly.');

    debug("press up");
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-01');
    shouldBeEqualToString('highlightedValue()', '1999-12');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-01,1999-12-02,1999-12-03,1999-12-04,1999-12-05,1999-12-06,1999-12-07,1999-12-08,1999-12-09,1999-12-10,1999-12-11,1999-12-12,1999-12-13,1999-12-14,1999-12-15,1999-12-16,1999-12-17,1999-12-18,1999-12-19,1999-12-20,1999-12-21,1999-12-22,1999-12-23,1999-12-24,1999-12-25,1999-12-26,1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31');

    debug("press down");
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31');
    shouldBeEqualToString('highlightedValue()', '2000-01');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31');
    
    debug("press down");
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-30,2000-01-31');
    shouldBeEqualToString('highlightedValue()', '2000-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29');

    debug("press left");
    eventSender.keyDown('leftArrow');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31');
    shouldBeEqualToString('highlightedValue()', '2000-01');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31');

    debug("press right");
    eventSender.keyDown('rightArrow');
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-30,2000-01-31');
    shouldBeEqualToString('highlightedValue()', '2000-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29');

    debug("press m");
    eventSender.keyDown('m');
    shouldBeEqualToString('currentMonth()', '2000-03');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2000-03');
    shouldBeEqualToString('highlightedDayCells()', '2000-03-01,2000-03-02,2000-03-03,2000-03-04,2000-03-05,2000-03-06,2000-03-07,2000-03-08,2000-03-09,2000-03-10,2000-03-11,2000-03-12,2000-03-13,2000-03-14,2000-03-15,2000-03-16,2000-03-17,2000-03-18,2000-03-19,2000-03-20,2000-03-21,2000-03-22,2000-03-23,2000-03-24,2000-03-25,2000-03-26,2000-03-27,2000-03-28,2000-03-29,2000-03-30,2000-03-31');

    debug("press shift + m");
    eventSender.keyDown('M');
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-30,2000-01-31');
    shouldBeEqualToString('highlightedValue()', '2000-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29');

    debug("press y");
    eventSender.keyDown('y');
    shouldBeEqualToString('currentMonth()', '2001-02');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2001-02');
    shouldBeEqualToString('highlightedDayCells()', '2001-02-01,2001-02-02,2001-02-03,2001-02-04,2001-02-05,2001-02-06,2001-02-07,2001-02-08,2001-02-09,2001-02-10,2001-02-11,2001-02-12,2001-02-13,2001-02-14,2001-02-15,2001-02-16,2001-02-17,2001-02-18,2001-02-19,2001-02-20,2001-02-21,2001-02-22,2001-02-23,2001-02-24,2001-02-25,2001-02-26,2001-02-27,2001-02-28');

    debug("press shift + y");
    eventSender.keyDown('Y');
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-30,2000-01-31');
    shouldBeEqualToString('highlightedValue()', '2000-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29');

    debug("press d");
    eventSender.keyDown('d');
    shouldBeEqualToString('currentMonth()', '2010-02');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2010-02');
    shouldBeEqualToString('highlightedDayCells()', '2010-02-01,2010-02-02,2010-02-03,2010-02-04,2010-02-05,2010-02-06,2010-02-07,2010-02-08,2010-02-09,2010-02-10,2010-02-11,2010-02-12,2010-02-13,2010-02-14,2010-02-15,2010-02-16,2010-02-17,2010-02-18,2010-02-19,2010-02-20,2010-02-21,2010-02-22,2010-02-23,2010-02-24,2010-02-25,2010-02-26,2010-02-27,2010-02-28');

    debug("press shift + d");
    eventSender.keyDown('D');
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-01');
    shouldBeEqualToString('selectedDayCells()', '2000-01-30,2000-01-31');
    shouldBeEqualToString('highlightedValue()', '2000-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29');

    debug("press enter");
    eventSender.keyDown('\n');
    shouldBeEqualToString('currentMonth()', '2000-02');
    shouldBeEqualToString('selectedValue()', '2000-02');
    shouldBeEqualToString('selectedDayCells()', '2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29');
    shouldBeEqualToString('highlightedValue()', '2000-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29');
    shouldBe('document.getElementById("month").value', '"2000-02"');

    // Test twice in case midnight.
    shouldBeTrue('testToday() || testToday()');

    debug('Check if tabbing works.');
    shouldBe('focusedElement()', '".list-view calendar-table-view"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".calendar-navigation-button"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".calendar-navigation-button today-button"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".calendar-navigation-button"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".month-popup-button"');

    debug("press esc");
    eventSender.keyDown('\x1B');
    waitUntilClosing(test2AfterClosing);
}

function test2AfterClosing() {
    shouldBeNull('document.getElementById("mock-page-popup")');

    finishJSTest();
}

function testToday() {
    eventSender.keyDown('t');
    var now = new Date();
    var expectedMonth = new popupWindow.Month(now.getFullYear(), now.getMonth()).toString();
    return selectedValue() === expectedMonth && highlightedValue() === expectedMonth && currentMonth() === expectedMonth;
}

function focusedElement() {
    var element = popupWindow.document.activeElement;
    if (!element)
        return null;
    var identifier = "." + element.className;
    if (element.value)
        identifier += "[value=" + element.value + "]";
    return identifier;
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
